Refactor top level definitions for Scala 3 compatibility
