+++ core/src/main/scala-2/anorm/SqlQueryCompat.scala
@@ -4,4 +4,4 @@
 
 package anorm
 
-trait AnormCompatSpec { _self: tests.AnormSpec => }
+private[anorm] trait SqlQueryCompat {}
